Rename items package to item
