+++ src/main/java/tracko/logic/commands/item/EditItemCommand.java
@@ -16,10 +16,10 @@ import tracko.logic.commands.CommandResult;
 import tracko.logic.commands.exceptions.CommandException;
 import tracko.logic.parser.CliSyntax;
 import tracko.model.Model;
-import tracko.model.items.Description;
-import tracko.model.items.Item;
-import tracko.model.items.ItemName;
-import tracko.model.items.Quantity;
+import tracko.model.item.Description;
+import tracko.model.item.Item;
+import tracko.model.item.ItemName;
+import tracko.model.item.Quantity;
 import tracko.model.tag.Tag;
 
 /**
