+++ src/main/java/tracko/logic/parser/ParserUtil.java
@@ -9,9 +9,9 @@ import java.util.Set;
 import tracko.commons.core.index.Index;
 import tracko.commons.util.StringUtil;
 import tracko.logic.parser.exceptions.ParseException;
-import tracko.model.items.Description;
-import tracko.model.items.ItemName;
-import tracko.model.items.Quantity;
+import tracko.model.item.Description;
+import tracko.model.item.ItemName;
+import tracko.model.item.Quantity;
 import tracko.model.order.Address;
 import tracko.model.order.Email;
 import tracko.model.order.Name;
