+++ src/main/java/tracko/logic/parser/item/AddItemCommandParser.java
@@ -15,10 +15,10 @@ import tracko.logic.parser.Parser;
 import tracko.logic.parser.ParserUtil;
 import tracko.logic.parser.Prefix;
 import tracko.logic.parser.exceptions.ParseException;
-import tracko.model.items.Description;
-import tracko.model.items.Item;
-import tracko.model.items.ItemName;
-import tracko.model.items.Quantity;
+import tracko.model.item.Description;
+import tracko.model.item.Item;
+import tracko.model.item.ItemName;
+import tracko.model.item.Quantity;
 
 /**
  * Parses input arguments and creates a new/update AddItemCommand Object.
