+++ src/main/java/tracko/model/item/exceptions/DuplicateItemException.java
@@ -1,4 +1,4 @@
-package tracko.model.items.exceptions;
+package tracko.model.item.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Items
