+++ src/main/java/tracko/ui/ItemCard.java
@@ -7,7 +7,7 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import javafx.scene.text.TextAlignment;
-import tracko.model.items.Item;
+import tracko.model.item.Item;
 
 /**
  * A UI component that displays information of an {@code Item}.
